Re-order fields on default ticket
This moves the QR code higher up on the ticket, so it doesn't end up in
the previous place which typically is right on the fold of the paper.
Only the locations of fields is changed.

It would still definitely be beneficial if somebody with better design
skills did a once-over on the tickets to make them  look better.

Fixes #7