Add support for uploaded bank files
This adds generic support to all managed bank accounts for handling
uploaded files. The idea is that for banks that don't have a working API
to download transaction list, one can be downloaded manually from the
bank and then uploaded into the system, and thereby using the same type
of matching etc.

Each account also gets to set a time limit after which it shows up as
"you should be uploading a file!" in the interface, and a scheduled job
will email a reminder to the treasurer email daily (can be set to 0 to
turn off).