Fix permissions checks for documentation pages
Always require a login for viewing pages. This was already
"meta-enforced" by crashing when the user wasn't logged in, but should
do a proper redirect. We always want to enforce this login so we don't
end up with google indexing the help pages of every site for example.

Second, allow talk voters and any members of any of the global
permission groups to view the documentation. The comment already covered
the talkvoters, just not the implementation, but since we have
documentation for things outside of confreg now, also include
permissions on those parts of the system.

Partial work on #19