Add Generic Managed Bank Accounts
This is a subset of the managed bank accounts to handle the case where
no API or scrapers exist, but there is a "reasonable format" available
for downloadable transactions. For these, the file format definition can
go in a JSON file instead of requiring specific code for each banks
format.

This centralizes a bit of the code written specificaly for the pgeu bank
and kept in a separate repository, and makes it more configurable to
suit the regular case.

In doing so, also create the concept of a BankStatementRow that contains
the actual output from the bank files. This was previously handled
independently by each django app for different payment providers, making
it hard to provide a good generic view.

And finally, also create generic views to see this set of bank statement
rows coming out of the files, so it's possible to follow things up. For
now this can be filtered by file (which file did these transactions come
from) or just by payment method ("all transactions for this bank
account"). This can be extende din the future to cover more advanced
filters and searching.

Include generic parsers for Credit Mutuel (the one previously living in
code form in the pgeu repository) in France, SEB and Swedbank (both in Sweden).