Show speaker the last notified status instead of current status
When batch-approving talks during voting, there was a state where the
speaker could see the state as being "pending approval" (for example),
while the organisers were still working on the details and had not yet
sent out notifications. And could even confirm it.

Instead, base the output views on the actual *notified* status of the
talk. This way, the notifications act as a gateway both on the email and
the web site, making all talks have their status announced at the same
time.

Fixes #38