Fix highlighting of status column
Commit 1225151dc001abe4a41fead744655c534630e898 added an <a> to the
status column making it better highlighted. Unfortunately that only
fixed the initial view of the page, and broke immediately as one changed
the status of a talk, since this would remove the <a> element again.

Fix this by changing the contents of the <a> element instead of
replacing it, when the status is updated.

Fixes #33 again