Dynamically add/remove button for pending notifications in vote page
Commit b76f1601443e29178f9523bebe7b9823e6aee4ed moved the pending
notifications alert to the top of the page, making them more noticable,
but it only updated them on page load and not when status changes were
made.

This commit makes it update them on every status change, so changes are
immediately visible. Doing so with the previous section-with-headline
would make the whole page jump when the status was changed which is far
from friendly, so replace the section with a highlighted (bootstrap
warning style) button on the same row as the filter buttons.

This also gets the vertical space back, for quicker access to the
voting. And since it's not a separate section anymore, also remove the
headline for Voting, saving even more space.

Finally, this changes the voting API to return JSON instead of
delimited-text, also making it more future-proof.

Fixes #36