Move volunteer schedule back into admin framework
Instead of rendering the volunteer schedule as part of the conference
templates, render them as part of the backend.

The volunteer schedule does a number of things that makes it put a lot
of requirements on the templates around it, and that's currently built
around bootstrap. This makes really hard to use in some conferences, and
since it's only for volunteers anyway, making it render in the plainer
framework doesn't have a big downside.

This also makes it integreate nicer with the backend templates for the
administrators, which is an improvement.