Redo the volunteer schedule management to be more dynamic
This moves most of the interaction over to javascript and API calls,
which makes it a lot nicer to interact with when the volunteer schedules
are large, since it doesn't require a complete reload of the page every
time.

In passing also change around how the header looks by removing
unnecessary headers and replacing mutiple paragraphs of short text with
buttons intead.

Fixes #23