Replace datetime.now() with timezone.now()
As a step on the way to better timezone support, use the django function
timezone.now() instead of datetime.now(). As long as we haven't enabled
timezones globally this becomes a no-op and does exactly what it did
before, but once timezones are enabled it will generate datetimes that
are aware of this.

No functionality change but gets a lot of boiler-plate out of the way
making the verification of the rest of the timezone work easier.