Track canceled registrations
Instead of deleting a registration when it's canceled, track when it was
canceled in a new field, canceledat. This makes it possible to keep
track of who the speakers are and potentially email them in the future.

There are two main things that need to get done that are not included in
this commit:
1. Reports need to be adjused. Some are, but all need to be summarized.
   In particular, time reports need to be properly taught about it and
   probably given the option to actually graph it.
2. Cross-conference email need to be given the option to include or not
   include canceled registrations.