Add basic registration activity log
This adds a basic log to each registration of activity that happens on
it, from creation until possible canceling.

More details can be added as we go, but tihs is a good start.

One big improvement needed is to track exactly *which* fields changed
and to/from what when making changes to a registration (that would help
with support cases).