Fix deletion of unpaid registrations
Tracking of canceled registration broke the deletion of registrations
that were *not* paid, because it tried to flag it as canceled even
though not paid. As we don't care about the history of registrations
that have never been paid, just delete the record completely when this
happens instead (same as it worked before we started tracking
cancellations).