Don't send removal notices to regs added by third parties
If a registration was never confirmed, the attendee may not know they
were ever added (because someone may have just done it in error), and
would thus just get confused by the message, so don't send it.

If the attendee added themselves, then keep sending the notifications of
course, and likewise if they were ever confirmed.