Add support for bulk-cancel of registrations
This adds the ability to mark multiple registrations in the registration
list and cancel them all at once, reducing the number of required
clicks.

All registrations being canceled must choose the same refund policy.
This means that refund policies that have a percentage other than 0 or
100 can lead to "interesting" results, but that is up to the
administrator to review at each individual case.

The bulk cancelation will be rolled back if any individual cancelation
fails. However, once a refund is scheduled the registration is canceled
(just like before), meaning that if the refund itself fails later, the
registration is already canceled. But since we now have tracking of
canceled registrations, this is something that can be properly followed
up (manually).