Raise exception when processing refunds if any fail
Now that we don't crash on the individual refund in order to proceed
with others, it became a bit too easy to miss the fact that something
went wrong at all. So instead, collect the error state until the end and
raise an exception if anything went wrong earlier, thereby triggering an
error flag in the job scheduler.