Clean up cross conference email js/css loading
Move the JS off to it's own file as one should, move inline CSS to the
actual stylesheet, and the stylesheet to the CSS file.

This adds an API endpoint to get the conferences that the user has
permissions on, instead of stitching it into the middle of the JS with
the template.