Implement native pages for conference news articles
This means each conference news now gets a proper permanent link
going to /events/<conference>/news/<slug>-<id>/, and this is the link
that's put out in the RSS feeds.

A page is created at /events/<conference>/news/ which contains the full
news history for the conference.

This creates a new "section" in the templates called "news", for those
skins that care about the section.

Fixes #44