Add basic support for canceling and refunding sponsorships
This adds support for both canceling a sponsorship (with or without
refund) and for refunding a sponsorship invoice (with or without
cancel).

There are more manual steps than in cancellations of registrations (for
example, no refund patterns), but refunding of sponsorships is a much
less common occurrance (there are fewer sponsors), so that seemed like
overengineering.