Remove date parameter to create_accounting_entry
This should always be set to todays date, so set it in the function
instaed.

This changes the behaviour of the paypal plugin to actually use todays
date rather than a date parsed out from the notification. In theory this
could change the accounting entry date in case the notification is so
delayed it ends up on a different day, but either one could be argued to
be equally correct in this case, and using the date of creation of the
transaction makes tracking things a lot easier.

While at it, clean up some imports that have been missed earlier.