Add a field for timezone name to each conference
This field is not currently *used* for anything, but it will be in the
future once we get real and proper timezone support into the system.

When the migration is run it tries to figure out the timezone of the
conferece based on the location. (It does *not* try to match it with the
existing timediff field, because that whole idea is so terrible) If this
fails, it wills et the timezone to the default timezone configured in
the system.

A list of which timezones were matched to what, and which failed, is
sent to DEFAULT_EMAIL at the end of the migration. It is then expected
that the operator goes through and verifies the timezones of all
conferences before proceeding to the migration that has the actual
timezone support which comes later.

No changes to the actual time values stored in the database are done at
this point, this adjustment will have to happen once the timezone
mapping has been manually corrected.