Move admin injection code into specific app
Instead of the even more hackish solution to use an app called _initial
with an init module, move the code to inject the concurrency protection
into the app in util/, and instead make sure this app is loaded *before*
the django.contrib.admin.

Also move the check for svgcairo and qrencode to the util app, because
that's cleaner.

This removes the _initial app completely.