Add protection against accidentally over-refunding an invoice
This check is normally done higher up in the stack to make sure we can
give a proper error message. But if the upper end of the stack failed to
do it's checking, add a hard exception in case an attempt is made to
refund more of an invoice than is actually available.