Add support for refunding prepaid voucher orders
If a prepaid voucher order was purchased through an invoice, add the
ability to refund it. Refunding can only happen on orders that have no
vouchers used, so when refunding an order that has been used, one must
first cancel all the registrations on it, and then refund the order
itself. Only full refunds of the order are currently supported.