Make sure name in TW sending has at least two parts
Apparently it's forbidden to create recipients with only one name, there
has to be at least two parts separated by a space (but can be more than
two parts). If this happens, just duplicate the name with a space in
between, as the name of the recipient is actually completely irrelevent.