Look for invoices to refund every 1 hour instead of 4
We have a very cheap check to see if there is antyhing at all to do, so
we can definitely run it a bit more often to decrease waiting times. We
still want to batch things out a bit like this due to how some of the
upstream APIs work, but once an hour is definitely still within reason.