Enforce the timezone is reset between requests
Turns out django does not reset the timezone on requests, unless it's
done explicitly. So add a new middleware that does exactly this -- and
then individual requests will set it back to the conference timezone as
required.

Not doing this could lead to very interesting results in a
multi-threaded server...