Enforce that manually confirmed regs are not paid
A manually confirmed registration goes directly to confirmed when, well,
confirmed, and can thus not receive an invoice. Enforce this property by
refusing to create registration types that use this special type if they
also have a cost, and update the documentation to be even more clear
about this.

We may in the future want to add a registration type that has a manual
validation step but otherwise follows the flow of regular registrations
and payments, but that's not this function.