Generate pending bank matchers for managed bank accounts
For managed bank accounts that don't have support for automated refunds,
we can still generate pending bank matcher for a reasonable string,
since the transaction will eventually show up. If it doesn't, or amounts
don't match etc, the operator can always drop the pending bank matcher
as well. This also means it can complete the accounting entry for
automatic closing.

In this also update the email that's sent to treasurer to suggest the
proper string.

Fixes #47