Add ability to re-issue a sponsorship invoice
This will cancel the old invoice and create a new one with a new invoice
number, but it will do so without canceling the sponsorship itself, and
thus without causing issues with existing claimed benefits and other
sponsorship details. The new invoice will copy the due date of the old
invoice.

Fixes #12