Track sent conference emails in the database
This adds a tracking similar to how we were already tracking attendee
and sponsor emails, but for the cross-conference emails.

Store a disconnected set of rules and recipients (just email
addresses), so that the record of who the email was sent to is kept even
if the recipient is removed (for example by having their registration
canceled).

In the future this can be enhanged with support for VERP to track
individual emails that are bouncing and set automatic opt-out, but
that's not included yet.

Closes #20