Add URL field to rooms
This makes it possible to link each room to for example a map pointing
out where the room is, or a stream link for a virtual conference.

In passing, change it so that a room isn't listed at all if it's not set
(instead of showing a "Room:" entry with no contents) on the session
details page.

Suggested by Dan Langille

Fixes #49