Use sitebase for links in ical feed
For some historic and unknown reason, the template for generating the
ical feeds used a separate variable for server name and constructed the
URL from that. Probably because it predates the SITEBASE variables being
added, and was missed when those were created. So change it to use the
SITEBASE, like all other links everywhere else.

Issue spotted by Dan Langille