Add support for an admin to add/edit/delete slides on a session
In particular being able to delete slides if bad ones are uploaded is
important, but it's also useful for the admin to be able to upload
slides that are for example emailed in from speakers for some reason
unable to use the system.

Both uploaded PDFs and copy/pasted links are supported.

Fixes #48