Make sender name for membership emails configurable
Previously, it was only possible to change the address in the membership
system, but the name was taken from the global settings. This seemed
inconsistent, so make it configurable at the membership system level,
using the global org name as the default.

In passing (and the main reason for working on this), make sure that the
*reminder* and *expiry* emails that are sent from the nightly job sets a
name at all. Previously only the manual emails would use the name, and
the reminders would just a blank email address without a name.