Ensure scheduled jobs runner quits on global exceptions
Most exceptions are handled by the runner itself, but if an exception
happened at the global level only the main thread would exit, leaving
the watcher thread running alone, meaning in reality the service hung.
Since the system is designed to have systemd (or whatever else is used)
automatically restart it on exit, this change ensures that the process
exits in this case triggering that behaviour.

This can happen for example if the database is restarted at the wrong
moment.