Create a Dockerfile to ease development
Additionally document in code the required setup and have an example of use. It
is by no means intended to be run in production.

The base image is from ubuntu 18.04 LTS which is the most recent version of
ubuntu that can match the python/pip/virtualenv requirements. For example ubuntu
20.04 will require Pillow 6.2 instead of 5.4.1 and virtualenv will have to be
launched without the -no-site-packages option.

The setup is based on the name pgeusystem which is used throughout, e.g. admin
name, db name etc.

Some basic instructions on how to build, run and use the image are added to the
README on the devsetup directory, which by itself is now listed in the root
README file to provide visibility.