Add support for web based meetings
Instead of requiring people to get on IRC, support member meetings in a
web browser. To make this work, there will be a simple websockets based
server (hosted in a separate repository) that will act as a relay,
and a trivial web app to handle the frontend.

Also include a native handling of polls in the system, including
timeouts, being typical actions during these meetings.

Meeting log and handling goes in the database, making it easy to extract
later for generation of official meeting minutes.

Code by me, layout and styling by Ilaria Battiston.