Implement conference JWT tokens
This allows an external site to integrate with the conference
registration system. All things around registration and possible
payments can be done on the regular site. An URL is then exposed where a
lgoged in user can get a token which in turn can be used by an external
app to request a signed JWT with information about the registration.