Ignore adyen reports in non-csv format
It seems we get a notification into the system for PDF reports manually
ordered in the interface as well (at least sometimes), which would then
crash the download process as we try to store the results in a text
field. Add a check for csv, and ignore all non-csv reports. We still
store them in the list of report (and just pre-flag them as processed),
so that we will have a record of which reports have been delivered if
something changes down the line.