Fix calculation of due date on sponsor invoices
The field on an invoice is named duedate, but actually contains a
timestamp -- this is very confusing and should be fixed at some point.
Because this means that passing an actual date into the duedate
parameter when creating an invoice will cause an exception because it
lacks timezone information.

This affected sponsorshp invoices if generated less than 30 day sbefore
the conference was to begin, voucher and discount code invoices.