Fix return codes in volunteer schedule API
All endpoints would return 500 errors regardless of what the actual
error was. This triggers django to send an email to the administrator.
Since these are actually normal things to happen, change it to send 4xx
errors (depending on the actual error) instead which is both
conceptually correct and gets rid of those emails.

The client side JS consuming the API treats 5xx and 4xx the same way, so
it makes no difference there.