Fix social posting window vs timezones
Automated twitter posts such as those coming from sponsor posts should
be adjusted against the configured time window on the conference, so
they don't post in the middle of the night.

If a post was entered after the end of the timezone window it should get
adjusted to the start time of the following day. However, if the post
was entered when it was *already* the following day in the conference
timezone (but not in UTC), it would incorrectly get adjusted to the
start of the *current* day (UTC), and thus post immediately.

Fix this by doing the adjustment on the time in the conference
timezone instead of UTC.