Silence warnings from send_queued_mail on django 3.2
Silence the deprecation warnings that otherwise happens on Django 3.2
about the deprecation of JSONField. This can be reverted once the full
migration to 3.2 is done, but since this is a cronjob that usually runs
every  minute or two it generates an insane amount of warnings until
that point, so just shut it up.