Centralize and fix calculation of VAT totals for registrations
The preview of VAT totals for conference registrations (and related
things like additional options) used a slightly different order of
calculation vs the invoice system, which could lead to differences in
the last decimal between the preview and the actual invoice.

Fix taht by centralizing the multiple different places that sued a
different method into one function, and change that function to work the
same way as the invoice system.