Re-implement Adyen payments using the Pay by Link system
The current Hosted Payment Pages (HPP) system is deprecated by Adyen and
will be removed in Oct 2022. So we replace the implementation with one
using the Payment Link system.

The big advantage of HPP vs the other options provided by Adyen is that
*all* processing is handled at Adyen, thereby not requiring any PCI
certification. And luckily the same applies for Payment Link, which is
why it's picked over the other available options.

The difference in functionality is that instead of encoding and signing
the payment information as part of the URL of the page, and API call is
made to Adyen to get a link (passing the information as a json object)
and the user is redirected to the return URL. Once that part is done,
the processesing is handled by Adyen. There is no longer an actual
authorization result as part of the return url (something HPP had but we
didn't use anyway), but as before we send the user in a refresh loop
while we wait for the asynchronous notification to arrive.

Payment Link also providers a generally more modern interface and works
better on mobile for example, which while not necessary very important
to our customers, is a nice touch. It may also open the door for some
other payment methods in the future.