Make it possible for a "require claim" sponsorship benefit to auto-confirm
Benefits set to "require explicit claiming" that are always going to be
confirmed but where the claiming step is still needed to know that the
sponsor actually wants the benefit can now be configured to
automatically confirm as soon as they are claimed.

For cases where somethign has to be checked before confirmation this is
of course still possible, and remains the default.