Fix format of timestamps when extracted from DateTimeRange
We want these to use the configured datetime format and not fallback to a
system default (which includes timezone specifications, and we don't
want that since a single event is only in one timezone).

To make the view even easier to parse, if both the start and the end is
on the same day (which is normally is), just show the date once, and the
interval between the two times.

Fixes #69