Separate internal notes from submission notes on call for papers
Move the "copied from <previous conference>" note over to "internal
notes" so the submitter cannot remove/change the information, and also
add to it what the status of the talk was in the conference it was
copied from.

This clearly separate out submission notes (by the speaker) from
internal notes (by the system, or by organizers).

Internal notes can *only* be seen by the conference staff and is bisible
when editing a conference in the backend, or when voting on talks. It is
never shown to the end user.

Note that this is a session global note -- it's not the same as the
notes that cna be made by each individual reviewer during talk voting.

Fixes #26