Fix opt-out handling for registrations with no accounts
When making a registration for somebody else, it is possible to create a
registration which is not connected to an account. The attendee will be
encouraged to attach it to an account and told how to do it, but if they
don't they remain in the system with no account connection. With the
existing system for opt-out, this would cause a crash because all
opt-out is based on account id.

For this purpose, add a `localoptout` field to the registration which
allows this individual registration to opt out of communications. It
will only do that one registration, and will not be able to do things
like opt out of a complete conference series.

This can only happen for registrations and not for speakers, as speakers
always have an account -- so stick to using the accuont there.