Track sent membership emails in the database
This keeps track of emails that have been sent, for future references.
For each email we track the contents and metadata about the email sent,
as well as exactly which members it was sent to. We do not track the
individual email addresses of the member, just the membership, so if
they later change their email, it does not affect the list of emails,
 but it also means the individual address a mail was sent to cannot be
found in the system later (that's for the mail server logs to handle).

While doing this, also implement a shortcut button for "send email to
all active members", instead of having to go throuhg the path of
filtering out the members, selecting them all, and sending email. The
ability to send individual email is kept, but the "send to all active
members" is likely to be the most used method.

Fixes #55