Don't crash hard on invalid values in attendee report filters
For example putting a non-date into a date field should at least give a
better error message.

For HTML reports we can just remove the filter and run the rest, with a
warning. For other formats, that don't allow us to add warnings, just
return an error message and don't run the report at al.