Fix interaction between remove_field() and fieldsets in backend forms
Commit 0d5dde299 put a workaround in place for this bug with accounting
objects, but it turns out the actual bug goes deeper and is in the
interaction between BackendForm.remove_field() and the use of fieldsets.

Basically, for each backend process, if one user opened a form (for
example a backend registration form) on a conference where some fields
were removed (e.g. tshirt size) then this would be persisted and anybody
who later opened a form for a different conference with this field
enabled would get a crash.

To fix this, move the fieldsets list into the Meta class, and if it's
present in the Meta class then overwrite it in the main form class every
time the form is instantiated (usign a deep copy, so that removed fields
come back).

Do this conditionally, so that any forms that have fieldsets defined on
the main form but not in Meta will continue to work. As long as they
don't use remove_field() they will continue to work.

Update the only two forms that used both remove_field() and fieldsets,
being BackendConferenceRegistrationForm and BackendSuperConferenceForm
to have fieldsets in Meta.