Require 2+ options in conference feedback questions
It makes no sense to have an input with just one possible option. Zero
is valid as it makes it a freetext field, or 2 or more, but not one.
This can for example happen when the wrong separator of options are
used, to best to be blocked - and also try to specifically note the case
when it's : instead of ; as a separator and highlight the problem.

"Discovered" by Daniel Gustafsson