Remove an attendee from signups when canceled
When a registration is canceled, rmeove it from any signups it has
responded to (and log that). It is not possible to add a canceled
registraiton, and it's not included in the list of registrations
awaiting signup, but it would still be included in the summary of
numbers and responses, which makes no sense.

Fixes #75