Make call for papers and sponsors controllable by a time range
Allows specifying at which time call for papers and call for sponsors
are opened, instead of manually having to flip a checkbox. For the call
to be opened, it needs to *both* satisfy the checkbox and the time
range, unless the timerange is left completely empty (the default) i
which case the checkbox is used along like before.

Fixes #34