Fix error msg when signing up as sponsor and conf is not open
At some point during the auth refactoring this turned into a generic
Access Denied message, which is not very user friendly. There should be
no way to get to these links when it's not open from within the system,
but it definitely can be linked from the outside so we should have a
good message.

Fixes #11