Use proper timezones in volunteer twitter
Set the default timezone based no the conference timezone offset (even
if javascript clearly hates the idea that a timezone other than that of
the curernt browser session even exists) since that's what's actually
interesting. And indicate exactly which timezone is used for the "last
refreshed" data since that may or may not be different depending on the
browser.