Fix speaker profile creation without call for papers
When arriving directly at the speaker profile form, the user would be
presented with a 404 error instead of a form to create a new profile.
The only way to create a new profile was to first submit a talk, and
*then* go edit the profile.

The reason for this was that the code was catching the
Speaker.DoesNotExist exception, but it was using get_object_or_404() to
look up the speaker, which throws a completely different exception. Fix
this by directly invoking Speaker.objects.get().

Fixes #85