Streamline speaker profile setup for new speakers
Instead of automatically creating an empty profile for new speakers when
they submit (and subsequently having to remind them to fix it), replace
the "new submission" paragraph with instructions to create the profile
*first*. Once the profile is created there will be a link back to make a
submission (this link was alraedy there for those that entered that way,
but that path was broken due to #85) at which point the new submission
link will properly function.

Hopefully this will reduce the number of empty submissions in the
system.